LinuxSDK 2020-11 (LinuxSDK_v2020.11.0.tar.gz)
DTAPI 5.43.0.147
Dta 4.28.1.271
DtaNw 3.5.11.42
DtPcie 1.11.0.262
Dtu 4.14.0.85

+++ LinuxSDK/Drivers/DtPcie/Source/DtBcSDITXPLL.c
@@ -279,7 +279,7 @@ DtStatus  DtBcSDITXPLL_C10A10_Calibrate(DtBcSDITXPLL* pBc)
         DtStringUInt64ToDtStringAppend(&Str, 10, pBc->m_pCore->m_pDevInfo->m_Serial);
         DtStringAppendChars(&Str, "] ");
         DtStringAppendChars(&Str, "SDI TX PLL calibration has been completed in: ");
-        DtStringUInt64ToDtStringAppend(&Str, 10, (TimeEnd-TimeStart)/1000);
+        DtStringUInt64ToDtStringAppend(&Str, 10, ((UInt32)(TimeEnd-TimeStart))/1000);
         DtStringAppendChars(&Str, "ms");
 
         DtEvtLogReport(&pBc->m_pCore->m_Device.m_EvtObject,
